move event traits out of a feature gate
